Flat tests
* Move unit-tests and helpers to signgle package
for the sake of simplisity
