+++ scripts/test.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
 
 echo "==> Running 'gotest' ..."
-for testingpkg in $(go list ./tests/.../); do
-  go test -count=1 -v $testingpkg
-done
+go test -count=1 -v ./gcore/.
